reorganize code (better file separation and cleaner import/using statementss)
