+++ src/backgrounds.jl
@@ -1,5 +1,3 @@
-export T1Decay, T2Dephasing, Depolarization, PauliNoise, AmplitudeDamping
-
 """A background describing the T₁ decay of a two-level system."""
 struct T1Decay <: AbstractBackground
     t1
