+++ src/baseops/apply.jl
@@ -1,5 +1,3 @@
-import QuantumInterface: apply!
-
 """
 Apply a given operation on the given set of register slots.
