+++ src/concurrentsim.jl
@@ -1,8 +1,3 @@
-using ResumableFunctions
-import ConcurrentSim # Should be using
-using ConcurrentSim: Environment, request, release, now, active_process, timeout, Store, @process, Process, put, get
-using Printf
-
 export @simlog, nongreedymultilock, spinlock, get_time_tracker
 
 macro simlog(env, msg) # this should be part of @process or @resumable and just convert @info and company
